Add files via upload
I've fixed some issues in these new versions. The code should be working with UD conllu files. I wasn't able to test it on CIEP data, that's why I didn't make changes in the CIEP scripts. The changes are as follows:
1. arguments_lemmas_forms_new.py:
1.1. A comment about languages with cases: sv, da and no can be excluded, in my view.
1.2. findArguments(): Lemmas now include morphological features, except for case. This is done to solve the problem with number and definiteness marking, which can interfere with case forms overlap.
1.3. cleanArguments(): A comment about Welsh: please check!
1.4. cleanArguments(); Added "al" to Spanish as a possible object marker. 
 

2. lemma_overlap_new.py:
count_overlap_lemmas(): removed the condition for nouns only, which interfered. I think it's not necessary.